+++ src/Driver/macros.py
@@ -31,25 +31,43 @@ logger.debug(f'Initializing {__file__}')
 #     from _linux_macros import perform_press
 
 
+if sys.platform == 'win32':
+    # make this a class
+    try:
+        # the keys are stored in the environment variables
+        SPEECH_KEY = os.environ.get('AZURE_SPEECH_KEY')
+        SPEECH_REGION = os.environ.get('AZURE_SPEECH_REGION')
+        if SPEECH_KEY is None or SPEECH_REGION is None:
+            raise ValueError("One of the AZURE_SPEECH_KEYS is not defined.", (SPEECH_KEY,SPEECH_REGION))
+        # This example requires environment variables named "SPEECH_KEY" and "SPEECH_REGION"
+        speech_config = speechsdk.SpeechConfig(subscription=SPEECH_KEY, region=SPEECH_REGION)
+        audio_config = speechsdk.audio.AudioOutputConfig(use_default_speaker=True)
+        # The language of the voice that speaks.
+        speech_config.speech_synthesis_voice_name='en-US-JennyNeural'
+        speech_synthesizer = speechsdk.SpeechSynthesizer(speech_config=speech_config, audio_config=audio_config)
 
-# make this a class
-try:
-    # the keys are stored in the environment variables
-    SPEECH_KEY = os.environ.get('AZURE_SPEECH_KEY')
-    SPEECH_REGION = os.environ.get('AZURE_SPEECH_REGION')
-    if SPEECH_KEY is None or SPEECH_REGION is None:
-        raise ValueError("One of the AZURE_SPEECH_KEYS is not defined.", (SPEECH_KEY,SPEECH_REGION))
-    # This example requires environment variables named "SPEECH_KEY" and "SPEECH_REGION"
-    speech_config = speechsdk.SpeechConfig(subscription=SPEECH_KEY, region=SPEECH_REGION)
-    audio_config = speechsdk.audio.AudioOutputConfig(use_default_speaker=True)
-    # The language of the voice that speaks.
-    speech_config.speech_synthesis_voice_name='en-US-JennyNeural'
-    speech_synthesizer = speechsdk.SpeechSynthesizer(speech_config=speech_config, audio_config=audio_config)
+    except Exception as e:
+        logger.error(f'Error setting up speech config: {e}')
+        #toaster.show_toast("Speech Config Error", "See Log File for more information", duration=5, threaded=True)
+        send_notification(title='Speech Config Error', msg='Speech Config Error\nSee Log File for more information', expire_time=2, urgency='normal')
+    
+    
+    def textToSpeech():
+        logger.info("in textToSpeech")
+        """Microsoft Azure moves to TLS 1.2 On October 31 2024
+        IDK what that really means, so if this breaks after that date, we will know why
+        https://learn.microsoft.com/en-us/security/engineering/solving-tls1-problem
+        """
+        perform_hotkey(['ctrl', 'c'])
+        time.sleep(0.1)
+        text = pyperclip.paste()
+        speech_synthesizer.speak_text(text)
+        #speech_synthesis_result = speech_synthesizer.speak_text_async(text)
 
-except Exception as e:
-    logger.error(f'Error setting up speech config: {e}')
-    #toaster.show_toast("Speech Config Error", "See Log File for more information", duration=5, threaded=True)
-    send_notification(title='Speech Config Error', msg='Speech Config Error\nSee Log File for more information', expire_time=2, urgency='normal')
+    def stopSpeech():
+        logger.info("in stopSpeech")
+        speech_synthesizer.stop_speaking()
+        #speech_synthesizer.stop_speaking_async()
 
 
 # timer function
@@ -76,22 +94,6 @@ def get_time(func):
 """
 These should work on both systems
 """
-def textToSpeech():
-    logger.info("in textToSpeech")
-    """Microsoft Azure moves to TLS 1.2 On October 31 2024
-    IDK what that really means, so if this breaks after that date, we will know why
-    https://learn.microsoft.com/en-us/security/engineering/solving-tls1-problem
-    """
-    perform_hotkey(['ctrl', 'c'])
-    time.sleep(0.1)
-    text = pyperclip.paste()
-    speech_synthesizer.speak_text(text)
-    #speech_synthesis_result = speech_synthesizer.speak_text_async(text)
-
-def stopSpeech():
-    logger.info("in stopSpeech")
-    speech_synthesizer.stop_speaking()
-    #speech_synthesizer.stop_speaking_async()
 
 def libreOffice_zoomin():   
     pyautogui.keyDown('ctrl')
@@ -194,7 +196,8 @@ def perform_press(key):
     logger.debug(f"perform_press {key = }")
     platformModule._perform_press(key)
 
-def change_desktop(direction:str, focused_app=None) -> None: #change desktop hotkey, where direction is either 'left' or 'right'. Will alt+tab if specific program is focused
+#def change_desktop(direction:str, focused_app=None): #change desktop hotkey, where direction is either 'left' or 'right'. Will alt+tab if specific program is focused
+def change_desktop(direction, focused_app): #change desktop hotkey, where direction is either 'left' or 'right'. Will alt+tab if specific program is focused
     """Changes the desktop, moves left or right
 
     Args:
@@ -205,7 +208,7 @@ def change_desktop(direction:str, focused_app=None) -> None: #change desktop hot
     
     platformModule._change_desktop(direction, focused_app)
 
-def moveAppAccrossDesktops(hwnd: int, movement: str) -> int:
+def moveAppAccrossDesktops(app_id: int, move: str):
     """Moves an app accross desktops.
     This function requires the VirtualDesktopAccessor.dll on windows,
     It is not implemented yet on linux
@@ -216,13 +219,14 @@ def moveAppAccrossDesktops(hwnd: int, movement: str) -> int:
     'Current': Moves to the current desktop
     
     Args:
-        hwnd (HWND): The HWND(handle) of the app
+        app_id (HWND|window_id): The HWND(handle) of the app
         movement (str): Movement Type
 
     Returns:
         bool: It will return 1 if succesfull and 0 if not
     """
-    return platformModule._moveAppAccrossDesktops(hwnd, movement)
+    #return platformModule._moveAppAccrossDesktops(hwnd, movement)
+    return platformModule._moveAppAccrossDesktops(app_id=app_id, movement= move)
 
 def Image_to_text2():
     """Presses Win Shift S to open snippet mode, waits for mouse release then does tesseract OCR
@@ -233,7 +237,12 @@ def Image_to_text2():
     logger.debug("Image_to_text2 has finished")
 
 
-
+def get_focused():
+    
+    return platformModule._get_focused()
+def get_focused_name():
+    
+    return platformModule._get_focused_name()
 
 
 # import SpotifyController